microservice-user/requirements.txt
Flask==3.0.0
redis==5.0.1
PyJWT==2.8.0
werkzeug==3.0.1
python-dotenv==1.0.0
bcrypt==4.1.2
pytest == 8.4.0
pytest-cov == 4.1